move Apps components into apps folder
