+++ src/pages/Apps/AppCard.tsx
@@ -1,15 +1,14 @@
 import Tooltip from "rc-tooltip";
 import { FC, useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
-import { ReactComponent as LinkIcon } from "../../../assets/external-link.svg";
-import { ReactComponent as InfoIcon } from "../../../assets/information-circle.svg";
-import { ReactComponent as LockIcon } from "../../../assets/lock-open.svg";
-import { ReactComponent as PlusIcon } from "../../../assets/plus.svg";
-import AppIcon from "../../../container/AppIcon/AppIcon";
-import { AppStatus, AuthMethod } from "../../../models/app-status";
-import { App } from "../../../models/app.model";
-
-import ButtonWithSpinner from "../../Shared/ButtonWithSpinner/ButtonWithSpinner";
+import AppIcon from "../..//container/AppIcon/AppIcon";
+import { AppStatus, AuthMethod } from "../..//models/app-status";
+import { App } from "../..//models/app.model";
+import { ReactComponent as LinkIcon } from "../../assets/external-link.svg";
+import { ReactComponent as InfoIcon } from "../../assets/information-circle.svg";
+import { ReactComponent as LockIcon } from "../../assets/lock-open.svg";
+import { ReactComponent as PlusIcon } from "../../assets/plus.svg";
+import ButtonWithSpinner from "../../components/Shared/ButtonWithSpinner/ButtonWithSpinner";
 
 export type Props = {
   appInfo: App;
